[Dev tool] Clean doc for artifact-path
